Initial commit: advent25 solution + advent24 day1
Initial commit for the advent25 solution. gitignore boilerplate is for VS in general, and will surely need some work for JGrasp if Java is used in the project.
Use Advent 2024 Day 1 as a hello-world for the toolsets and a warmup for new developers on the team. Work items are all TODO in comments